libc pass-2: generator fixes (Kotlin-keyword escaping, octal underscores, unsigned const literals, union field nullability)
Regenerated 234 generated files. Raw error count rises because fixing parse failures unmasks previously-hidden unresolved-reference errors; the dominant blockers are architectural (native-only types in commonMain, no canonical primitives, expect/actual not yet applied).

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/bsd/freebsdlike/freebsd/freebsd13/X8664.kt
@@ -2,9 +2,10 @@
 package io.github.kotlinmania.libc.unix.bsd.freebsdlike.freebsd.freebsd13
 
 import io.github.kotlinmania.libc.*
+import kotlinx.cinterop.COpaquePointer
 
 public const val PROC_KPTI_CTL: CInt = PROC_PROCCTL_MD_MIN
 public const val PROC_KPTI_CTL_ENABLE_ON_EXEC: CInt = 1
 public const val PROC_KPTI_CTL_DISABLE_ON_EXEC: CInt = 2
 public const val PROC_KPTI_STATUS: CInt = PROC_PROCCTL_MD_MIN + 1
-public const val PROC_KPTI_STATUS_ACTIVE: CInt = 0x80000000.toInt()
+public const val PROC_KPTI_STATUS_ACTIVE: CInt = 0x80000000
